+++ src/pages/BorrowerProfile/FundedBorrowerProfile.vue
@@ -221,21 +221,13 @@ const newFundedBorrowerPageExpKey = 'new_funded_borrower_page';
 export default {
 	name: 'FundedBorrowerProfile',
 	metaInfo() {
-		const canonicalUrl = `https://${this.$appConfig.host}${this.$route.path}`.replace('funded', 'lend');
 		return {
-			title: `${this.loan?.name ?? ''} From ${this.loan.geocode?.country?.name ?? ''}'s loan has been funded`,
+			title: `${this.loan.name} from ${this.loan.geocode?.country?.name}'s loan has been funded!`,
 			meta: [
 				{
 					vmid: 'description',
 					name: 'description',
-					content: `A loan helped ${this.loan?.use ?? ''}`,
-				}
-			],
-			link: [
-				{
-					vmid: 'canonical',
-					rel: 'canonical',
-					href: canonicalUrl
+					content: `A loan helped ${this.loan.use}`,
 				}
 			]
 		};