+++ CHANGELOG.md
@@ -1,3 +1,86 @@
+### [2.216.1](https://github.com/kiva/ui/compare/v2.216.0...v2.216.1) (2022-07-05)
+
+
+### Bug Fixes
+
+* bugs related to FLSS loan channel ([d6b072c](https://github.com/kiva/ui/commit/d6b072cd66000e6148980aabc15f7e45bcca0e26))
+
+## [2.216.0](https://github.com/kiva/ui/compare/v2.215.0...v2.216.0) (2022-07-05)
+
+
+### Features
+
+* apply location to query params filtering and shared state ([7eda2a9](https://github.com/kiva/ui/commit/7eda2a9869dbbf8a02e78c747a8bbe6ba42d8984))
+
+
+### Bug Fixes
+
+* missing story for active ISO codes ([08bac1f](https://github.com/kiva/ui/commit/08bac1fe6b0f7c86a3ca2916d5f9849a8370abcc))
+
+## [2.215.0](https://github.com/kiva/ui/compare/v2.214.2...v2.215.0) (2022-07-05)
+
+
+### Features
+
+* use canvas pool to reduce live-loan memory usage ([2e474d8](https://github.com/kiva/ui/commit/2e474d834a28cbd627fa4456068262b2d07b2b28))
+
+### [2.214.2](https://github.com/kiva/ui/compare/v2.214.1...v2.214.2) (2022-07-05)
+
+
+### Bug Fixes
+
+* use automatic handling of memcached connection issues ([879ac88](https://github.com/kiva/ui/commit/879ac88558b185e441769be00681b202cfc9df49))
+
+### [2.214.1](https://github.com/kiva/ui/compare/v2.214.0...v2.214.1) (2022-07-01)
+
+
+### Bug Fixes
+
+* missing punctuation ([d1271c9](https://github.com/kiva/ui/commit/d1271c90f4ad979b4efc6a3fdd7a5b325e0cfbd6))
+
+## [2.214.0](https://github.com/kiva/ui/compare/v2.213.2...v2.214.0) (2022-07-01)
+
+
+### Features
+
+* distribution model to FLSS channel mapping ([7097c4d](https://github.com/kiva/ui/commit/7097c4d986ed6a284a83c183659810cb70a318b3))
+* experiment switching loan channel data to FLSS ([0b8f365](https://github.com/kiva/ui/commit/0b8f36586958eeb380507b8bffbffa67c36a32bd))
+* loan channel FLSS experiment tracking ([c95ab1a](https://github.com/kiva/ui/commit/c95ab1aaa03585462bf483334466cc84d4f57af6))
+
+
+### Bug Fixes
+
+* added defensive code and cleaned up code consistency ([1ff9883](https://github.com/kiva/ui/commit/1ff988332a1dc71c8192679b0c369f5d913ac541))
+* matched the promise chain pattern ([da39b8f](https://github.com/kiva/ui/commit/da39b8f6583802d7b009cc329bacddb401b6c902))
+
+### [2.213.2](https://github.com/kiva/ui/compare/v2.213.1...v2.213.2) (2022-06-30)
+
+
+### Bug Fixes
+
+* update copy for borrower profile share card ([2033dad](https://github.com/kiva/ui/commit/2033dad7c3fb477f3984b1d806a9708f7fd5909f))
+
+### [2.213.1](https://github.com/kiva/ui/compare/v2.213.0...v2.213.1) (2022-06-30)
+
+
+### Bug Fixes
+
+* add inviter name to shareLink func ([5b112d2](https://github.com/kiva/ui/commit/5b112d2e986c2953591e788ea66d627a2ad0f60f))
+* canonical loop issue ([cd1b224](https://github.com/kiva/ui/commit/cd1b22434fe60935c744dff55e70c026631a4ee0))
+* resolve conflicts ([cf92a18](https://github.com/kiva/ui/commit/cf92a18d5d37153f6fb7e34a8ba84d586f8afe83))
+
+## [2.213.0](https://github.com/kiva/ui/compare/v2.212.1...v2.213.0) (2022-06-30)
+
+
+### Features
+
+* recommended by lender category, category page ([dae992f](https://github.com/kiva/ui/commit/dae992f32fdf0e9f0238ede07af54384792712f5))
+
+
+### Bug Fixes
+
+* moved recommended by lenders category page to new route to avoid duplicate prefetch ([b4c255c](https://github.com/kiva/ui/commit/b4c255c10f4149b1f4bd3e363b92d8a6b84f11b4))
+
 ### [2.212.1](https://github.com/kiva/ui/compare/v2.212.0...v2.212.1) (2022-06-30)